Apply the bash script to all server files
